# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: AliasDeck Server API
  version: "0.3.0"
  description: >
    The /api/v1 control-plane surface exposed by `aliasdeck serve`: CRUD for
    aliases, profiles and devices; operator session auth; single-use
    enrollment tokens for device registration; device token rotation and
    revocation. This document is embedded in the server binary
    (internal/api/openapi.go) and served at GET /api/v1/openapi.yaml; a
    bidirectional coverage test (internal/api/openapi_coverage_test.go)
    fails if this file and the router's own route table ever disagree.
servers:
  - url: /api/v1
security:
  - sessionToken: []
paths:
...
  /api/v1/devices:
    get:
      summary: List every device
      responses:
        "200":
          description: Every device.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/devices/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      summary: Get one device
      responses:
        "200":
          description: The device.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Rename a device or change its profile membership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Device"
      responses:
        "200":
          description: The updated device.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidReference"
    delete:
      summary: Delete a device
      responses:
        "204":
          description: The device was deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/devices/{id}/revoke:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      summary: Revoke a device and every device-kind token it holds
      responses:
        "204":
          description: The device and its tokens were revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/devices/{id}/token:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      summary: Revoke a device's existing token(s) and issue a new one
      responses:
        "200":
          description: The device's id and its new device token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: An operator session token, minted by POST /api/v1/auth/login.
    enrollmentToken:
      type: http
      scheme: bearer
      description: A single-use enrollment token, minted by POST /api/v1/enrollment-tokens.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
...
    Unauthorized:
      description: No valid session (or, for device registration, enrollment token) was presented.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No resource with that id exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InvalidReference:
      description: The write names another resource (a profile or device id) that does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    DeviceToken:
      type: object
      properties:
        deviceId:
          type: string
        deviceToken:
          type: string
          description: The device's wire token, `add_<lookup>.<secret>`. Shown once.
...
    Device:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        name:
          type: string
        hostname:
          type: string
        platform:
          type: string
          enum: [macos, linux, windows]
        shell:
          type: string
          enum: [zsh, bash, powershell]
        architecture:
          type: string
        profileIds:
          type: array
          items:
            type: string
        lastSeenAt:
          type: string
          format: date-time
          readOnly: true
        lastSyncAt:
          type: string
          format: date-time
          readOnly: true
        clientVersion:
          type: string